Fixed the last dart score logic to ensure that the last dart must be a successful checkout of a double or bullseye

+++ DartsMathsBlazorGame/Components/DartsMaths.razor.cs
@@ -57,7 +57,7 @@ public partial class DartsMaths
 
     private void ClearGuess()
     {
-        SelectedScoreArea = ScoreArea.Single;
+        SelectedScoreArea = ScoreArea.Double;
         ScoreValue = 1;
         IsGuessCorrect = null;
     }
